% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_fecrt.R
\name{analyse_fecrt}
\alias{analyse_fecrt}
\title{analyse_fecrt}
\usage{
analyse_fecrt(data_1, data_2, H0_I, H0_A, paired, tail = 0.025,
  mean_ratio = 1, conjugate_priors = c(0, 0), delta = 1,
  beta_iters = 10^4, approx = 1, dobson_cl = c(0.005, 0.995),
  dobson_priors = c(1, 1))
}
\arguments{
\item{dobson_priors}{}
}
\description{
This function analyses a given dataset using different methods
}
\examples{
data1 <- rnbinom(10, mu=20, size=1)
data2 <- rnbinom(10, mu=2, size=0.75)
analyse_fecrt(data1, data2, H0_I=0.95, H0_A=0.9, paired=FALSE)
}
