% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_analysis.R
\name{count_analysis}
\alias{count_analysis}
\alias{count.analysis}
\alias{fec.analysis}
\alias{FEC.analysis}
\alias{bayescount.single}
\title{Count data analysis}
\usage{
count_analysis(data = stop("Data must be specified"), model = "ZILP",
  alt.prior = FALSE, adjust.zi.mean = FALSE, raw.output = FALSE,
  likelihood = FALSE, ...)
}
\description{
Analyses count data for difference to either known or fixed mean using any of (1) Christian approach, (2) my approach, (3) bootstrap, (4) MCMC (default 1-3).  (4) calls count_model and then runs the model using runjags, and is the only one to permit ZI and LP.  Requires Just Another Gibbs Sampler (JAGS) (see http://mcmc-jags.sourceforge.net)
}
\seealso{
\code{\link{count_model}}, \code{\link{reduction_analysis}}, \code{\link{bayescount}}, \code{\link[runjags]{autoextend.jags}}, \code{\link[runjags]{runjags}}
}
