% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_model.R
\name{count_model}
\alias{count_model}
\alias{count.model}
\alias{fec.model}
\alias{FEC.model}
\alias{run.model}
\title{Count data model}
\usage{
count_model(data = stop("No data supplied"),
  model = stop("No model specified"), call.jags = FALSE,
  alt.prior = FALSE, monitor.lambda = FALSE, monitor.deviance = FALSE,
  ...)
}
\description{
Creates a count model for the specified data and returns it as an un-run runjags object
}
\seealso{
\code{\link{count_analysis}}, \code{\link{reduction_model}}, \code{\link{bayescount}}, \code{\link[runjags]{autoextend.jags}}, \code{\link[runjags]{runjags}}
}
