% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_power.R
\name{count_power}
\alias{count_power}
\alias{count.power}
\alias{fec.power}
\alias{FEC.power}
\title{Count data power calculations}
\usage{
count_power(type = "fec", cutoff = if ("fec") 400 else 95,
  alternative = "two.sided", conf.level = 0.05, maxiterations = 1e+06,
  mc.decimals = 2, mc.conf = 0.99, feedback = FALSE, ...)
}
\description{
Power calculation based on BNB code and christian approach yet to be written.  Do either test vs known single mean, or test for difference in mean to observed dataset (k can vary between datasets or be fixed).
}
\seealso{
\code{\link{count_analysis}}, \code{\link{reduction_power}}, \code{\link{bayescount}}
}
