% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_precision.R
\name{count_precision}
\alias{count_precision}
\alias{count.precision}
\alias{fec.precision}
\alias{FEC.precision}
\alias{fec.power.limits}
\alias{FEC.power.limits}
\title{Count data precision calculations}
\usage{
count_precision(meanepg = 200, g.faeces = 3, sensitivity = 1/25,
  replicates = 1, animals = 10, coeffvarmcm = 0.1, coeffvarrep = 0.4,
  coeffvarind = 0.3, coeffvargroup = 0.7, true.sample = FALSE,
  type = "confidence", precision = if (type == "confidence") 0.1 else NA,
  lower.interval = meanepg * (1 - precision), upper.interval = meanepg * (1
  + precision), confidence = 0.95, iterations = if (type == "confidence")
  10^6 else 10^5, mc.decimals = 2, mc.conf = 0.99, feedback = FALSE,
  forcesim = FALSE)
}
\description{
Precision analysis by Monte Carlo simulation and examination of distribution of mean
}
\seealso{
\code{\link{count_analysis}}, \code{\link{count_power}}, \code{\link{bayescount}}
}
