% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_typologies.R
\name{efficacy_typologies}
\alias{efficacy_typologies}
\title{Obtain typology classifications relating to a theoretical dataset that may be encountered in practice}
\usage{
efficacy_typologies(
  sum,
  N,
  k = c(1, 1),
  cor = 0.1,
  paired = TRUE,
  T_I = 0.99,
  T_A = 0.95,
  R = c(1, 1),
  S = c(1, 1),
  method = "all",
  alpha = 0.025,
  bnb_priors = c(0, 0),
  use_delta = NA,
  beta_iters = 10^4,
  binomial_priors = c(1, 1),
  binomial_cl_adj = 0.2
)
}
\arguments{
\item{sum}{a length-2 integer representing the total count in the control/pre-treatment data and the treatment/post-treatment data}

\item{N}{a length-2 integer representing the number of individuals in the control/pre-treatment data and the treatment/post-treatment data}

\item{k}{a length-2 strictly positive continuous number giving the over-dispersion of the control/pre-treatment and treatment/post-treatment data samples}

\item{cor}{a length-1 number between 0 and 1 giving the correlation between pre- and post-treatment data (ignored for an unpaired study)}

\item{paired}{logical flag for a paired or unpaired study design assumption on which to simulate (and analyse) the data}

\item{T_I}{the threshold for inferioirty (target efficacy of the intervention)}

\item{T_A}{the threshold for non-inferioirty (target efficacy of the intervention minus a non-inferioirty margin delta)}

\item{R}{a length-2 integer giving the number of replicates used for control/pre-treatment and treatment/post-treatment individuals}

\item{S}{a length-2 numeric variable giving the counting sensitivity used for control/pre-treatment and treatment/post-treatment individuals}

\item{method}{a character vector indicating the statistical methods to show results for, where 'all' means all available methods}

\item{alpha}{the significance level to use for the classification, where 0.025 corresponds to 95\\% CI and 0.05 corresponds to 90\\% CI}

\item{bnb_priors}{priors to use for the BNB method}

\item{use_delta}{logical flag to use the delta method approximation for the BNB method (NA means to use it unless it fails)}

\item{beta_iters}{number of iterations to use for the Monte Carlo approximation of the beta distribution transformation for the BNB method (when use_delta==FALSE)}

\item{binomial_priors}{the priors to use for the Binomial method}

\item{binomial_cl_adj}{the adjustment in confidence level to use for the Binomial method relative to the others (the default corresponds to e.g. 99\\% rather than 95\\% CI) - where 0 > adj >= 1}
}
\value{
A data frame as would be returned by \code{\link{efficacy_analysis}} for a theoretical dataset with the given properties
}
\description{
This function provides the typology classifications (for each statistical method chosen) relating to a theoretical dataset with the given properties
}
\examples{
efficacy_typologies(sum=c(20,1), N=c(10,10), k=c(1,1), cor=0.25,
paired = TRUE, T_I = 0.99, T_A = 0.95)
}
\seealso{
\code{\link{bayescount}}, \code{\link{efficacy_frequencies}}
}
