% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k.R
\name{estimate_k}
\alias{estimate_k}
\title{Title}
\usage{
estimate_k(
  Nrange = c(5, 100),
  EPGrange = c(50, 1000),
  EDT = c(1, 2.5, 5, 10, 25, 50),
  target = 0.99,
  margin = c(0.04, 0.09),
  k1 = 1,
  k2 = 0.6,
  cor = 0.3,
  Nlength = 50,
  mulength = 20,
  EPGlength = 100,
  iters = 1000L
)
}
\arguments{
\item{cor}{}
}
\value{

}
\description{
Title
}
