% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecrt_sample_size.R
\name{fecrt_sample_size}
\alias{fecrt_sample_size}
\title{Title}
\usage{
fecrt_sample_size(
  Nrange = c(5, 100),
  EPGrange = c(50, 1000),
  EDT = c(1, 2.5, 5, 10, 25, 50),
  target = 0.99,
  margin = c(0.04, 0.09),
  kpre = 1,
  kpost = 0.6,
  cor = 0.3,
  Nlength = 100,
  EPGlength = 100
)
}
\arguments{
\item{cor}{}
}
\value{

}
\description{
Title
}
