% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch an interactive Shiny app contained within the bayescount package}
\usage{
launch_shiny(appname, ...)
}
\arguments{
\item{appname}{the name of an app to launch (see the details section)}

\item{...}{further arguments to be passed to \code{\link[shiny]{runApp}}}
}
\description{
This function provides a convinient way in which to launch one of the interactive shiny apps contained within this package, as detailed at http://www.fecrt.com/framework
}
\details{
The apps that are currently available are:
\itemize{
\item \strong{data_analysis} : analyse an existing dataset
\item \strong{typologies} : examine the typology resulting from a hypothetical dataset and implications of different interpretative frameworks for ERR / FECR
}
\itemize{
\item \strong{study_planning} : sample size calculations for a prospective study
}
}
\seealso{
\code{\link{bayescount}}
}
