% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersim.R
\name{powersim_paired}
\alias{powersim_paired}
\title{Title}
\usage{
powersim_paired(rvs = seq(0, 0.2, by = 0.001), N = 20, mu = 20,
  k1 = 1, k2 = 0.7, kc = 1.2, ta = 0.9, ti = 0.95,
  iters = 10^3, approx = 1, priors = c(0, 0), useml = FALSE,
  dobson_cl = c(0.005, 0.995), dobson_priors = c(1, 1), pm = c("BNB",
  "Levecke", "MLE", "WAAVP", "Dobson"), plot = TRUE)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
