% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersim.R
\name{powersim_unpaired}
\alias{powersim_unpaired}
\title{Title}
\usage{
powersim_unpaired(rvs = seq(0, 0.2, by = 0.001), N = 20, mu = 20,
  k1 = 1, k2 = 0.7, ta = 0.9, ti = 0.95, iters = 10^3,
  approx = 1, priors = c(0, 0), useml = FALSE, pm = c("BNB",
  "Levecke", "MLE", "WAAVP"), plot = TRUE)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
