% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_power.R
\name{reduction_power}
\alias{reduction_power}
\alias{reduction_pval}
\alias{reduction.power}
\alias{fecrt.power}
\alias{FECRT.power}
\title{Count reduction (eg FECRT) data power calculations}
\usage{
reduction_pval(pre_data, post_data, H0_1 = 0.9, H0_2 = 0.95, edt_pre = 1,
  edt_post = 1, prob_priors = c(1, 1), k_change = 1, true_k = NA,
  delta_method = TRUE, beta_iters = 10^6)
}
\description{
Power calculation based on BNB code and Christian approximation.
}
\seealso{
\code{\link{count_analysis}}, \code{\link{reduction_power}}, \code{\link{bayescount}}
}
