% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_precision.R
\name{reduction_precision}
\alias{reduction_precision}
\alias{reduction.precision}
\alias{fecrt.precision}
\alias{FECRT.precision}
\alias{fecrt.power.limits}
\alias{FECRT.power.limits}
\title{Count data reduction precision calculations}
\usage{
reduction_precision(meanepg = 200, reduction = 95, g.faeces = 3,
  sensitivity = 1/25, replicates = 1, animals = 10,
  pre.coeffvarrep = 0.4, pre.coeffvarind = 0.3, pre.coeffvargroup = 0.7,
  post.coeffvarrep = 0.4, post.coeffvarind = 0.3,
  post.coeffvargroup = 0.7, true.sample = FALSE, lower.interval = NA,
  upper.interval = NA, iterations = 1e+05, mc.decimals = 0.95,
  mc.conf = 0.99, feedback = FALSE, mc.output = FALSE)
}
\description{
Precision analysis by Monte Carlo simulation and examination of distribution of mean
}
\seealso{
\code{\link{reduction_analysis}}, \code{\link{reduction_power}}, \code{\link{bayescount}}
}
