% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescount-package.R
\docType{package}
\name{bayescount-package}
\alias{bayescount}
\alias{bayescount-package}
\alias{bayescountpackage}
\title{bayescount: Statistical Analyses and Sample Size Calculations for Over-dispersed Count Data}
\description{
A set of functions and interactive Shiny apps to allow analysis of
    over-dispersed count data such as that associated with egg reduction rate
    (ERR) and faecal egg count reduction test (FECRT) studies in parasitology.
    Functions and apps are also provided to assist in sample size calculations
    for future studies.  A working installation of JAGS is required for
    MCMC-based methods (see <http://mcmc-jags.sourceforge.net>).

\emph{Data analysis}

See the analyse_fecrt function

\emph{Sample size calculations}

TODO

\emph{Publication results}

Code is provided to replicate the simulation results in section 6 of "A hypothesis testing framework for the ratio of means of two negative binomial distributions" via the "htfsim" vignette:  vignette("htfsim", package="bayescount")
}
\references{
M. J. Denwood, S. W. J. Reid, S. Love, M. K. Nielsen, L. Matthews, I. J. McKendrick, and G. T. Innocent. Comparison of three alternative methods for analysis of equine Faecal Egg Count Reduction Test data. Prev. Vet. Med. (2009), doi:10.1016/j.prevetmed.2009.11.009

Denwood, M. J. (2010). A quantitative approach to improving the analysis of faecal worm egg count data. University of Glasgow. Retrieved from http://www.gla.ac.uk/media/media_149338_en.pdf
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ku-awdc/bayescount}
}


\code{\link{launch_shiny}}, \code{\link{analyse_fecrt}}
}
\author{
\strong{Maintainer}: Matthew Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Bob Wheeler (Copyright holder of the code in /src/dists.*) [copyright holder]
  \item Richard Brent (Original copyright holder of the code in /src/optimize.cc) [copyright holder]
}

}
\keyword{internal}
