% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k.R
\name{estimate_k}
\alias{estimate_k}
\title{Title}
\usage{
estimate_k(
  data_1,
  data_2,
  paired = TRUE,
  k_ratio = 1,
  min = 0.001,
  max = 20,
  tol = 0.01
)
}
\arguments{
\item{cor}{}
}
\description{
Title
}
