\name{fecrt}
\alias{fecrt}
\alias{FECRT}
\title{PERFORM AN ANALYSIS OF FAECAL EGG COUNT REDUCTION TEST DATA USING MCMC TO GIVE A PROBABILITY DISTRIBUTION OF VALUES FOR THE MEAN EGG COUNT REDUCTION}
\description{
   This function is currently included as a placeholder for a function currently in development.  Requires Just Another Gibbs Sampler (JAGS).  *THIS SOFTWARE IS INTENDED FOR EDUCATIONAL PURPOSES ONLY AND SHOULD NOT BE RELIED UPON FOR REAL WORLD APPLICATIONS*
}
\usage{
fecrt(data)
}
\arguments{
   \item{data}{an existing integer vector containing the data.  No default.}
}
\value{None.
}
\seealso{
   \code{\link{bayescount}}
   \code{\link{bayescount.single}}
}

\author{Matthew Denwood \email{m.denwood@vet.gla.ac.uk} funded as part of the DEFRA VTRI project 0101.}

\examples{
\dontrun{
fecrt()
}
}
\keyword{models}